Extract some functions into base/fs.cpp
This is the start of a bigger move. Moving all `fs_` prefixed functions into their own file.

Cousin pr #10855
Helps with #10093
Replaced #10185
